Move stack manager & manage stack modification ops
