+++ divi/src/script/interpreter.cpp
@@ -14,7 +14,7 @@
 #include "script/script.h"
 #include "uint256.h"
 
-#include <memory>
+#include <script/StackManager.h>
 
 using namespace std;
 
@@ -262,259 +262,6 @@ bool OpcodeIsDisabled(const opcodetype& opcode)
 
 
 
-typedef std::vector<valtype> StackType;
-
-class ConditionalScopeStackManager
-{
-private:
-    std::vector<bool> conditionalScopeReadStatusStack;
-    bool conditionalScopeNeedsClosing = false;
-
-    void CheckConditionalScopeStatus()
-    {
-        conditionalScopeNeedsClosing = std::count(conditionalScopeReadStatusStack.begin(),conditionalScopeReadStatusStack.end(),false)!=0;
-    }
-public:
-    bool StackIsEmpty() const
-    {
-        return conditionalScopeReadStatusStack.empty();
-    }
-    bool ConditionalScopeNeedsClosing() const
-    {
-        return conditionalScopeNeedsClosing;
-    }
-
-    void OpenScope(bool conditionStatus)
-    {
-        conditionalScopeReadStatusStack.push_back(conditionStatus);
-        conditionalScopeNeedsClosing |= (conditionStatus==false);
-    }
-    bool TryElse()
-    {
-        if(conditionalScopeReadStatusStack.empty()) return false;
-        conditionalScopeReadStatusStack.back()= !conditionalScopeReadStatusStack.back();
-        CheckConditionalScopeStatus();
-        return true;
-    }
-    bool TryCLoseScope()
-    {
-        if(conditionalScopeReadStatusStack.empty()) return false;
-        conditionalScopeReadStatusStack.pop_back();
-        CheckConditionalScopeStatus();
-        return true;
-    }
-};
-
-
-struct StackOperator
-{
-protected:
-    StackType& stack_;
-    StackType& altstack_;
-    unsigned& flags_;
-    ConditionalScopeStackManager& conditionalManager_;
-public:
-    StackOperator(
-        StackType& stack, 
-        StackType& altstack, 
-        unsigned& flags,
-        ConditionalScopeStackManager& conditionalManager
-        ): stack_(stack)
-        , altstack_(altstack)
-        , flags_(flags)
-        , conditionalManager_(conditionalManager)
-    {
-    }
-
-    virtual bool operator()(opcodetype opcode, ScriptError* serror)
-    {
-        return set_error(serror, SCRIPT_ERR_BAD_OPCODE);
-    }
-};
-
-struct DisabledOp: public StackOperator
-{
-    DisabledOp(
-        StackType& stack, 
-        StackType& altstack, 
-        unsigned& flags,
-        ConditionalScopeStackManager& conditionalManager
-        ): StackOperator(stack,altstack,flags,conditionalManager)
-    {
-    }
-
-    virtual bool operator()(opcodetype opcode, ScriptError* serror) override
-    {
-        if (flags_ & SCRIPT_VERIFY_DISCOURAGE_UPGRADABLE_NOPS)
-        {
-            return set_error(serror, SCRIPT_ERR_DISCOURAGE_UPGRADABLE_NOPS);
-        }
-        return true;
-    }
-};
-
-struct PushValueOp: public StackOperator
-{
-    PushValueOp(
-        StackType& stack, 
-        StackType& altstack, 
-        unsigned& flags,
-        ConditionalScopeStackManager& conditionalManager
-        ): StackOperator(stack,altstack,flags,conditionalManager)
-    {}
-
-    virtual bool operator()(opcodetype opcode, ScriptError* serror) override
-    {
-        CScriptNum bn((int)opcode - (int)(OP_1 - 1));
-        stack_.push_back(bn.getvch());
-        return true;
-    }
-};
-
-struct ConditionalOp: public StackOperator
-{
-public:
-    ConditionalOp(
-        StackType& stack, 
-        StackType& altstack, 
-        unsigned& flags,
-        ConditionalScopeStackManager& conditionalManager
-        ): StackOperator(stack,altstack,flags,conditionalManager)
-    {}
-
-    virtual bool operator()(opcodetype opcode, ScriptError* serror) override
-    {
-        switch(opcode)
-        {
-            case OP_IF: case OP_NOTIF:
-            {
-                // <expression> if [statements] [else [statements]] endif
-                bool fValue = false;
-                if (!conditionalManager_.ConditionalScopeNeedsClosing())
-                {
-                    if (stack_.size() < 1)
-                        return set_error(serror, SCRIPT_ERR_UNBALANCED_CONDITIONAL);
-                    valtype& vch = stack_.back();
-                    fValue = CastToBool(vch);
-                    if (opcode == OP_NOTIF)
-                        fValue = !fValue;
-                    popstack(stack_);
-                }
-                conditionalManager_.OpenScope(fValue);
-            }
-            break;
-
-            case OP_ELSE:
-            {
-                if (!conditionalManager_.TryElse())
-                    return set_error(serror, SCRIPT_ERR_UNBALANCED_CONDITIONAL);
-            }
-            break;
-
-            case OP_ENDIF:
-            {
-                if (!conditionalManager_.TryCLoseScope())
-                    return set_error(serror, SCRIPT_ERR_UNBALANCED_CONDITIONAL);
-            }
-            break;
-            default:
-            break;
-        }
-        return true;
-    }
-
-};
-
-struct StackOperationManager
-{
-private:
-    const std::set<opcodetype> upgradableOpCodes = 
-        {OP_NOP1,OP_NOP2,OP_NOP3,OP_NOP4,OP_NOP5,OP_NOP6,OP_NOP7,OP_NOP8,OP_NOP9,OP_NOP10};
-    const std::set<opcodetype> simpleValueOpCodes = 
-        {OP_1NEGATE ,OP_1 ,OP_2 ,OP_3 , OP_4 , OP_5 , OP_6 , OP_7 , OP_8,
-        OP_9 ,OP_10 ,OP_11 , OP_12 , OP_13 , OP_14 , OP_15 , OP_16};
-    const std::set<opcodetype> conditionalOpCodes = {OP_IF, OP_NOTIF, OP_ELSE,OP_ENDIF};
-    
-    StackType& stack_;
-    StackType& altstack_;
-    unsigned flags_;
-    ConditionalScopeStackManager conditionalManager_;
-    StackOperator defaultOperation_;
-    std::map<opcodetype, StackOperator*> stackOperationMapping_;
-
-    std::shared_ptr<StackOperator> disableOp_;
-    std::shared_ptr<StackOperator> pushValueOp_;
-    std::shared_ptr<StackOperator> conditionalOp_;
-private:
-
-    void SetMappingForUpgradableOpCodes()
-    {
-        for(const opcodetype& opcode: upgradableOpCodes)
-        {
-            stackOperationMapping_.insert({opcode, disableOp_.get() });
-        }
-    }
-    void SetMappingForSimpleValueOpCodes()
-    {
-        for(const opcodetype& opcode: simpleValueOpCodes)
-        {
-            stackOperationMapping_.insert({opcode, pushValueOp_.get() });
-        }
-    }
-
-    void SetMappingForConditionalOpCodes()
-    {
-        for(const opcodetype& opcode: conditionalOpCodes)
-        {
-            stackOperationMapping_.insert({opcode, conditionalOp_.get() });
-        }
-    }
-public:
-    StackOperationManager(
-        StackType& stack,
-        StackType& altstack,
-        unsigned flags
-        ): stack_(stack)
-        , altstack_(altstack) 
-        , flags_(flags)
-        , conditionalManager_()
-        , defaultOperation_(stack,altstack,flags,conditionalManager_)
-        , stackOperationMapping_()
-        , disableOp_(std::make_shared<DisabledOp>(stack_,altstack_,flags_,conditionalManager_))
-        , pushValueOp_(std::make_shared<PushValueOp>(stack_,altstack_,flags_,conditionalManager_))
-        , conditionalOp_(std::make_shared<ConditionalOp>(stack_,altstack_,flags_,conditionalManager_))
-    {
-        InitMapping();
-    }
-
-    void InitMapping()
-    {
-        SetMappingForUpgradableOpCodes();
-        SetMappingForSimpleValueOpCodes();
-        SetMappingForConditionalOpCodes();
-    }
-
-    StackOperator* GetOp(opcodetype opcode)
-    {
-        auto it = stackOperationMapping_.find(opcode);
-        if(it != stackOperationMapping_.end())
-        {
-            return it->second;
-        }
-        return &defaultOperation_;
-    }
-    
-    bool ConditionalNeedsClosing() const
-    {
-        return conditionalManager_.ConditionalScopeNeedsClosing();
-    }
-    bool ConditionalScopeIsBalanced() const
-    {
-        return conditionalManager_.StackIsEmpty();
-    }
-};
-
-
 bool EvalScript(vector<vector<unsigned char> >& stack, const CScript& script, unsigned int flags, const BaseSignatureChecker& checker, ScriptError* serror)
 {
     static const CScriptNum bnZero(0);
@@ -606,208 +353,11 @@ bool EvalScript(vector<vector<unsigned char> >& stack, const CScript& script, un
                     //
                     // Stack ops
                     //
-                    case OP_TOALTSTACK:
+                    case OP_TOALTSTACK: case OP_FROMALTSTACK: case OP_2DROP: case OP_2DUP: case OP_3DUP: case OP_2OVER: case OP_2ROT:
+                    case OP_2SWAP: case OP_IFDUP: case OP_DEPTH: case OP_DROP: case OP_DUP: case OP_NIP: case OP_OVER: case OP_PICK: 
+                    case OP_ROLL: case OP_ROT: case OP_SWAP: case OP_TUCK: case OP_SIZE:
                     {
-                        if (stack.size() < 1)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        altstack.push_back(stacktop(-1));
-                        popstack(stack);
-                    }
-                    break;
-
-                    case OP_FROMALTSTACK:
-                    {
-                        if (altstack.size() < 1)
-                            return set_error(serror, SCRIPT_ERR_INVALID_ALTSTACK_OPERATION);
-                        stack.push_back(altstacktop(-1));
-                        popstack(altstack);
-                    }
-                    break;
-
-                    case OP_2DROP:
-                    {
-                        // (x1 x2 -- )
-                        if (stack.size() < 2)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        popstack(stack);
-                        popstack(stack);
-                    }
-                    break;
-
-                    case OP_2DUP:
-                    {
-                        // (x1 x2 -- x1 x2 x1 x2)
-                        if (stack.size() < 2)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        valtype vch1 = stacktop(-2);
-                        valtype vch2 = stacktop(-1);
-                        stack.push_back(vch1);
-                        stack.push_back(vch2);
-                    }
-                    break;
-
-                    case OP_3DUP:
-                    {
-                        // (x1 x2 x3 -- x1 x2 x3 x1 x2 x3)
-                        if (stack.size() < 3)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        valtype vch1 = stacktop(-3);
-                        valtype vch2 = stacktop(-2);
-                        valtype vch3 = stacktop(-1);
-                        stack.push_back(vch1);
-                        stack.push_back(vch2);
-                        stack.push_back(vch3);
-                    }
-                    break;
-
-                    case OP_2OVER:
-                    {
-                        // (x1 x2 x3 x4 -- x1 x2 x3 x4 x1 x2)
-                        if (stack.size() < 4)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        valtype vch1 = stacktop(-4);
-                        valtype vch2 = stacktop(-3);
-                        stack.push_back(vch1);
-                        stack.push_back(vch2);
-                    }
-                    break;
-
-                    case OP_2ROT:
-                    {
-                        // (x1 x2 x3 x4 x5 x6 -- x3 x4 x5 x6 x1 x2)
-                        if (stack.size() < 6)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        valtype vch1 = stacktop(-6);
-                        valtype vch2 = stacktop(-5);
-                        stack.erase(stack.end()-6, stack.end()-4);
-                        stack.push_back(vch1);
-                        stack.push_back(vch2);
-                    }
-                    break;
-
-                    case OP_2SWAP:
-                    {
-                        // (x1 x2 x3 x4 -- x3 x4 x1 x2)
-                        if (stack.size() < 4)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        swap(stacktop(-4), stacktop(-2));
-                        swap(stacktop(-3), stacktop(-1));
-                    }
-                    break;
-
-                    case OP_IFDUP:
-                    {
-                        // (x - 0 | x x)
-                        if (stack.size() < 1)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        valtype vch = stacktop(-1);
-                        if (CastToBool(vch))
-                            stack.push_back(vch);
-                    }
-                    break;
-
-                    case OP_DEPTH:
-                    {
-                        // -- stacksize
-                        CScriptNum bn(stack.size());
-                        stack.push_back(bn.getvch());
-                    }
-                    break;
-
-                    case OP_DROP:
-                    {
-                        // (x -- )
-                        if (stack.size() < 1)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        popstack(stack);
-                    }
-                    break;
-
-                    case OP_DUP:
-                    {
-                        // (x -- x x)
-                        if (stack.size() < 1)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        valtype vch = stacktop(-1);
-                        stack.push_back(vch);
-                    }
-                    break;
-
-                    case OP_NIP:
-                    {
-                        // (x1 x2 -- x2)
-                        if (stack.size() < 2)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        stack.erase(stack.end() - 2);
-                    }
-                    break;
-
-                    case OP_OVER:
-                    {
-                        // (x1 x2 -- x1 x2 x1)
-                        if (stack.size() < 2)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        valtype vch = stacktop(-2);
-                        stack.push_back(vch);
-                    }
-                    break;
-
-                    case OP_PICK: case OP_ROLL:
-                    {
-                        // (xn ... x2 x1 x0 n - xn ... x2 x1 x0 xn)
-                        // (xn ... x2 x1 x0 n - ... x2 x1 x0 xn)
-                        if (stack.size() < 2)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        int n = CScriptNum(stacktop(-1), fRequireMinimal).getint();
-                        popstack(stack);
-                        if (n < 0 || n >= (int)stack.size())
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        valtype vch = stacktop(-n-1);
-                        if (opcode == OP_ROLL)
-                            stack.erase(stack.end()-n-1);
-                        stack.push_back(vch);
-                    }
-                    break;
-
-                    case OP_ROT:
-                    {
-                        // (x1 x2 x3 -- x2 x3 x1)
-                        //  x2 x1 x3  after first swap
-                        //  x2 x3 x1  after second swap
-                        if (stack.size() < 3)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        swap(stacktop(-3), stacktop(-2));
-                        swap(stacktop(-2), stacktop(-1));
-                    }
-                    break;
-
-                    case OP_SWAP:
-                    {
-                        // (x1 x2 -- x2 x1)
-                        if (stack.size() < 2)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        swap(stacktop(-2), stacktop(-1));
-                    }
-                    break;
-
-                    case OP_TUCK:
-                    {
-                        // (x1 x2 -- x2 x1 x2)
-                        if (stack.size() < 2)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        valtype vch = stacktop(-1);
-                        stack.insert(stack.end()-2, vch);
-                    }
-                    break;
-
-
-                    case OP_SIZE:
-                    {
-                        // (in -- in size)
-                        if (stack.size() < 1)
-                            return set_error(serror, SCRIPT_ERR_INVALID_STACK_OPERATION);
-                        CScriptNum bn(stacktop(-1).size());
-                        stack.push_back(bn.getvch());
+                        if(!stackManager.GetOp(opcode)->operator()(opcode,serror)) return false;
                     }
                     break;
